#35: Extract GridPinService from Grid component
